validations and implements show to get sales
now client show has a filter for month and year

+++ app/controllers/clients_controller.ts
@@ -19,18 +19,61 @@ export default class ClientsController {
 
     return clients
   }
+  // detalhar um(a) cliente e vendas a ele(a) (show):
+  //         trazer as vendas mais recentes primeiro;
+  //         possibilidade de filtrar as vendas por mês + ano;
+  async show({ params, response, request }: HttpContext) {
+    const { month, year } = request.qs()
 
-  async show({ params }: HttpContext) {
-    // TODO: integrate with sales
-    return Client.findOrFail(params.id)
+    const client = Client.find(params.id)
+    if (!client) {
+      return response.status(404).json({ message: 'Client not found' })
+    }
+    if (month && year) {
+      if (month < 1 || month > 12) {
+        return response.status(400).json({ message: 'Invalid month' })
+      }
+
+      return Client.query()
+        .where('id', params.id)
+        .preload('sales', (salesQuery) => {
+          salesQuery
+            .select('id', 'quantity', 'total_price', 'created_at')
+            .whereRaw('extract(month from created_at) = ?', [month])
+            .whereRaw('extract(year from created_at) = ?', [year])
+            .orderBy('created_at', 'desc')
+        })
+        .first()
+    }
+    return Client.query()
+      .where('id', params.id)
+      .preload('sales', (salesQuery) => {
+        salesQuery
+          .select('id', 'quantity', 'total_price', 'created_at')
+          .orderBy('created_at', 'desc')
+      })
   }
 
   async store({ request, response }: HttpContext) {
-    // TODO: add validation
     const { client, phone, address } = request.all()
-    // if (!name || !cpf) {
-    //   throw new Error('Name and CPF are required')
-    // }
+    if (!client.name || !client.cpf) {
+      return response.status(400).json({ message: 'Name and CPF are required' })
+    }
+    if (!phone.phone) {
+      return response.status(400).json({ message: 'Phone is required' })
+    }
+    if (
+      !address.state ||
+      !address.city ||
+      !address.neighborhood ||
+      !address.street ||
+      !address.number ||
+      !address.zipCode
+    ) {
+      return response
+        .status(400)
+        .json({ message: 'State, city, neighborhood, street, number and zip code are required' })
+    }
 
     const clientReq = await Client.create({ ...client })
     const phoneReq = await Phone.create({ ...phone, clientId: clientReq.id })
@@ -39,32 +82,44 @@ export default class ClientsController {
     return response.status(201).json({ clientReq, phoneReq, addressReq })
   }
 
-  async update({ request, params }: HttpContext) {
+  async update({ request, params, response }: HttpContext) {
     const { client, phone, address } = request.all()
 
     try {
       if (client) {
-        const clientReq = await Client.findOrFail(params.id)
+        const clientReq = await Client.find(params.id)
+        if (!clientReq) {
+          return response.status(404).json({ message: 'Client not found' })
+        }
         clientReq.merge(client).save()
       }
       if (phone) {
         // NOTE: the phone needs to be improved for multiple phones
-        const phoneReq = await Phone.findByOrFail('clientId', params.id)
+        const phoneReq = await Phone.find('clientId', params.id)
+        if (!phoneReq) {
+          return response.status(404).json({ message: 'Phone not found' })
+        }
         phoneReq.merge(phone).save()
       }
       if (address) {
-        const addressReq = await Adress.findByOrFail('clientId', params.id)
+        const addressReq = await Adress.find('clientId', params.id)
+        if (!addressReq) {
+          return response.status(404).json({ message: 'Address not found' })
+        }
         addressReq.merge(address).save()
       }
 
-      return 'Client updated'
+      return response.status(200).json({ message: 'Client updated' })
     } catch (error) {
       throw new Error('Client not found')
     }
   }
 
-  async delete({ params }: HttpContext) {
-    const client = await Client.findOrFail(params.id)
+  async delete({ params, response }: HttpContext) {
+    const client = await Client.find(params.id)
+    if (!client) {
+      return response.status(404).json({ message: 'Client not found' })
+    }
     await client.delete()
     return 'Client deleted'
   }
